+++ asynctnt/__init__.py
@@ -2,7 +2,7 @@ from .connection import Connection, connect
 from .iproto.protocol import (
     Iterator, Response, TarantoolTuple, PushIterator,
     Schema, SchemaSpace, SchemaIndex, Metadata, Field,
-    Db
+    Db, IProtoError, IProtoErrorStackFrame
 )
 
 __version__ = '2.0.0b1'